System updates
+ Added clearing of atoms (delete all particles)
+ Added scaling charge signs (+/-) that get bigger when the camera moves farther away
! Fixed issue with particle list causing null reference and bad access exceptions
- Removed old/unused code from particle class

+++ Assets/ElementDesigner/FileSystem/Atom.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using System;
+using System.Linq;
 
 public enum BondType {Covalent, Ionic, Metallic}
 
@@ -11,8 +12,7 @@ public class Atom
     public int ProtonCount;
     public int NeutronCount;
     public int ElectronCount;
-
-    public int Charge;
+    public float Charge;
 
     // TODO: Valence shells and electrons divided into shells
     // TODO: Valence shell composition to affect reactivity
